+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## 0.0.3 (2016-06-13)
+
+Features:
+
+  - New "Interfaces" page; a framework for searching through interface data.
+
+Bugfixes:
+
+  - Interface.type can no longer return nil (@floored1585)
+
 ## 0.0.2 (2016-03-04)
 
 Features: